Convert langtons-ant
