--- sketches/langtons-ant/README.md
@@ -1,4 +1,0 @@
-# Langton's Ant
-
-A two-dimensional Turing machine.
-See <https://en.wikipedia.org/wiki/Langton%27s_ant>.
