--- sketches/langtons-ant/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="utf-8" />
-        <title>Langton's Ant</title>
-        <link rel="stylesheet" type="text/css" href="style.css" />
-        <script src="../../node_modules/p5/lib/p5.js"></script>
-    </head>
-    <body>
-        <script src="sketch.js"></script>
-    </body>
-</html>
