--- sketches/langtons-ant/style.css
@@ -1,12 +1,0 @@
-html,
-body {
-    height: 100%;
-}
-
-body {
-    background-color: #202020;
-    margin: 0;
-    display: flex;
-    justify-content: safe center;
-    align-items: safe center;
-}
