+++ src/app/app.routes.ts
@@ -1,6 +1,7 @@
 import { Routes } from "@angular/router";
 
 import { CollatzConjectureComponent } from "src/app/sketches/collatz-conjecture/collatz-conjecture.component";
+import { LangtonsAntComponent } from "src/app/sketches/langtons-ant/langtons-ant.component";
 import { WelcomeComponent } from "src/app/welcome/welcome.component";
 
 export const routes: Routes = [
@@ -8,10 +9,8 @@ export const routes: Routes = [
     {
         path: "sketch",
         children: [
-            {
-                path: "collatz-conjecture",
-                component: CollatzConjectureComponent,
-            },
+            { path: "collatz-conjecture", component: CollatzConjectureComponent },
+            { path: "langtons-ant", component: LangtonsAntComponent },
         ],
     },
 ];
