+++ src/app/welcome/welcome.component.html
@@ -1,6 +1,14 @@
 <h1>A collection of my sketches made with p5.js</h1>
 <dl>
     <dt>
+        <a routerLink="sketch/langtons-ant">Langton's Ant</a>
+    </dt>
+    <dd>
+        A two-dimensional Turing machine. See
+        <a href="https://en.wikipedia.org/wiki/Langton%27s_ant" target="_blank">Wikipedia</a>.
+    </dd>
+
+    <dt>
         <a routerLink="sketch/collatz-conjecture">collatz conjecture</a>
     </dt>
     <dd>
